Aorta_net/requirements.txt
absl-py                       2.1.0
acvl-utils                    0.2
ants                          0.0.7
asgiref                       3.7.2
batchgenerators               0.25
cachetools                    5.3.2
certifi                       2023.11.17
charset-normalizer            3.3.2
connected-components-3d       3.12.4
contextlib2                   21.6.0
contourpy                     1.2.0
cycler                        0.12.1
dicom2nifti                   2.4.10
Django                        4.2.10
dynamic-network-architectures 0.3.1
einops                        0.7.0
et-xmlfile                    1.1.0
filelock                      3.13.1
fonttools                     4.49.0
fsspec                        2023.12.2
future                        0.18.3
gevent                        24.2.1
google-auth                   2.27.0
google-auth-oauthlib          1.2.0
graphviz                      0.20.1
greenlet                      3.0.3
grpcio                        1.60.0
h5py                          3.10.0
huggingface-hub               0.20.3
idna                          3.6
imagecodecs                   2024.1.1
imageio                       2.33.1
importlib-metadata            7.0.1
importlib-resources           6.1.1
itk                           5.3.0
itk-core                      5.3.0
itk-filtering                 5.3.0
itk-io                        5.3.0
itk-numerics                  5.3.0
itk-registration              5.3.0
itk-segmentation              5.3.0
Jinja2                        3.1.3
joblib                        1.3.2
kiwisolver                    1.4.5
lazy_loader                   0.3
lightning-utilities           0.10.1
linecache2                    1.0.0
Markdown                      3.5.2
MarkupSafe                    2.1.4
matplotlib                    3.8.3
MedPy                         0.4.0
ml-collections                0.1.1
monai                         1.3.0
mpmath                        1.3.0
networkx                      3.2.1
nibabel                       5.2.0
nnunetv2                      2.2.1      /root/autodl-tmp/yml/project/p3/nnUNet
numpy                         1.26.3
nvidia-cublas-cu12            12.1.3.1
nvidia-cuda-cupti-cu12        12.1.105
nvidia-cuda-nvrtc-cu12        12.1.105
nvidia-cuda-runtime-cu12      12.1.105
nvidia-cudnn-cu12             8.9.2.26
nvidia-cufft-cu12             11.0.2.54
nvidia-curand-cu12            10.3.2.106
nvidia-cusolver-cu12          11.4.5.107
nvidia-cusparse-cu12          12.1.0.106
nvidia-nccl-cu12              2.18.1
nvidia-nvjitlink-cu12         12.3.101
nvidia-nvtx-cu12              12.1.105
oauthlib                      3.2.2
opencv-python                 4.9.0.80
openpyxl                      3.1.2
packaging                     23.2
pandas                        2.2.0
pillow                        10.2.0
pip                           23.3.1
protobuf                      4.23.4
pyarrow                       15.0.0
pyasn1                        0.5.1
pyasn1-modules                0.3.0
pydicom                       2.4.4
pyparsing                     3.1.1
python-dateutil               2.8.2
python-gdcm                   3.0.23
pytz                          2024.1
PyYAML                        6.0.1
requests                      2.31.0
requests-oauthlib             1.3.1
rsa                           4.9
safetensors                   0.4.2
scikit-image                  0.22.0
scikit-learn                  1.4.0
scipy                         1.12.0
seaborn                       0.13.2
setuptools                    68.2.2
SimpleITK                     2.3.1
six                           1.16.0
sqlparse                      0.4.4
sympy                         1.12
nvidia-nvjitlink-cu12         12.3.101
nvidia-nvtx-cu12              12.1.105
oauthlib                      3.2.2
opencv-python                 4.9.0.80
openpyxl                      3.1.2
packaging                     23.2
pandas                        2.2.0
pillow                        10.2.0
pip                           23.3.1
protobuf                      4.23.4
pyarrow                       15.0.0
pyasn1                        0.5.1
pyasn1-modules                0.3.0
pydicom                       2.4.4
pyparsing                     3.1.1
python-dateutil               2.8.2
python-gdcm                   3.0.23
pytz                          2024.1
PyYAML                        6.0.1
requests                      2.31.0
requests-oauthlib             1.3.1
rsa                           4.9safetensors                   0.4.2scikit-image                  0.22.0
scikit-learn                  1.4.0
scipy                         1.12.0
seaborn                       0.13.2
setuptools                    68.2.2
SimpleITK                     2.3.1
six                           1.16.0
sqlparse                      0.4.4
sympy                         1.12
tensorboard                   2.15.1
tensorboard-data-server       0.7.2
threadpoolctl                 3.2.0
tifffile                      2023.12.9
timm                          0.9.12
torch                         2.1.2
torchmetrics                  1.3.1
torchvision                   0.16.2
tqdm                          4.66.1
traceback2                    1.4.0
triton                        2.1.0
typing_extensions             4.9.0
tzdata                        2024.1
unittest2                     1.1.0
urllib3                       2.1.0
Werkzeug                      3.0.1
wheel                         0.41.2
yacs                          0.1.8
zipp                          3.17.0
zope.event                    5.0
